Scheduler utility support for ARINC-653 VM scheduler
Introduce support for scheduling all the DomUs that are assigned to
ARINC-653 scheduler operated cpupool. The initial attempt is to create
a non-preemptive scheduler that will schedule all the DomUs based on
the WCET(Worst Case Execution Time) assigned to them.

The utility also helps ensure the WCET for a given DomU does not exceed
its period. If otherwise, the utility reports an error and declines to
set the schedule for ARINC-653.

Besides, if the cumulative WCETs of all ARINC-operated DomUs fall short
of the hyperperiod, then idle time slices take over and run for the
remainder of the time-period until next major frame begins.

TODOs:
- The schedule generated is not yet delivered to the hypervisor via
  hypercall just yet. Once delivered, ARINC-653 will schedule the DomUs
  based on the schedule set by this utility.

- ARINC-653 scheduler does not support multi-vCPU implementation. Extend
  support for it later by bringing in Hierarchial Real time systems
  Partitioning - Regularity based Resource Partitioning Model (HiRTS-RRP)

Signed-off-by: Pavan Kumar Paluri <pvpk1994@gmail.com>