Support to set ARINC-653 non-preemptive schedule
Introduce support to successfully deliver the schedule to ARINC-653
DomU scheduler in the hypervisor. In order to make it work, the
timeslice in the utility has been adjusted such that each schedule
entry now runs for DEFAULT_TIMESLICE, i.e., timeslice duration
equivalent to that of hypervisor scheduler.

To be precise, the WCET, period and hyperperiod of the utility schedule
is updated to be in sync with exact duration of how it is supposed to
run in the hypervisor scheduler of ARINC-653.

Signed-off-by: Pavan Kumar Paluri <pvpk1994@gmail.com>